requirements.txt
numpy
pandas
h5py
scikit-learn
torch
torchvision
tensorflow-gpu==1.15.2
mia
advertorch
torchxrayvision